Move `EnvConfig` to the `habitat_common` crate
This is in to allow more widespread usage.
