+++ components/sup/src/config.rs
@@ -20,7 +20,6 @@
 //!
 //! See the [Config](struct.Config.html) struct for the specific options available.
 
-use std::env::VarError;
 use std::fmt;
 use std::io;
 use std::net::{IpAddr, Ipv4Addr, SocketAddr, SocketAddrV4, ToSocketAddrs};
@@ -29,71 +28,11 @@ use std::option;
 use std::result;
 use std::str::FromStr;
 
-use common::defaults::{GOSSIP_DEFAULT_IP, GOSSIP_DEFAULT_PORT, GOSSIP_LISTEN_ADDRESS_ENVVAR};
+use common::{
+    defaults::{GOSSIP_DEFAULT_IP, GOSSIP_DEFAULT_PORT, GOSSIP_LISTEN_ADDRESS_ENVVAR},
+    types::EnvConfig,
+};
 use error::{Result, SupError};
-use hcore::env as henv;
-
-/// Enable the creation of a value based on an environment variable
-/// that can be supplied at runtime by the user.
-pub trait EnvConfig: Default + FromStr {
-    /// The environment variable that will be parsed to create an
-    /// instance of `Self`.
-    const ENVVAR: &'static str;
-
-    /// Generate an instance of `Self` from the value of the
-    /// environment variable `Self::ENVVAR`.
-    ///
-    /// If the environment variable is present and not empty, its
-    /// value will be parsed as `Self`. If it cannot be parsed, or the
-    /// environment variable is not present, the default value of the
-    /// type will be given instead.
-    fn configured_value() -> Self {
-        match henv::var(Self::ENVVAR) {
-            Err(VarError::NotPresent) => Self::default(),
-            Ok(val) => match val.parse() {
-                Ok(parsed) => {
-                    Self::log_parsable(&val);
-                    parsed
-                }
-                Err(_) => {
-                    Self::log_unparsable(&val);
-                    Self::default()
-                }
-            },
-            Err(VarError::NotUnicode(nu)) => {
-                Self::log_unparsable(nu.to_string_lossy());
-                Self::default()
-            }
-        }
-    }
-
-    /// Overridable function for logging when an environment variable
-    /// value was found and was successfully parsed as a `Self`.
-    ///
-    /// By default, we log a message at the `warn` level.
-    fn log_parsable(env_value: &String) {
-        warn!(
-            "Found '{}' in environment; using value '{}'",
-            Self::ENVVAR,
-            env_value
-        );
-    }
-
-    /// Overridable function for logging when an environment variable
-    /// value was found and was _not_ successfully parsed as a `Self`.
-    ///
-    /// By default, we log a message at the `warn` level.
-    fn log_unparsable<S>(env_value: S)
-    where
-        S: AsRef<str>,
-    {
-        warn!(
-            "Found '{}' in environment, but value '{}' was unparsable; using default instead",
-            Self::ENVVAR,
-            env_value.as_ref()
-        );
-    }
-}
 
 #[derive(Clone, PartialEq, Eq, Debug)]
 pub struct GossipListenAddr(SocketAddr);
