+++ components/sup/src/manager/mod.rs
@@ -77,7 +77,8 @@ pub use self::sys::Sys;
 use self::user_config_watcher::UserConfigWatcher;
 use super::feat;
 use census::{CensusRing, CensusRingProxy};
-use config::{EnvConfig, GossipListenAddr};
+use common::types::EnvConfig;
+use config::GossipListenAddr;
 use ctl_gateway::{self, CtlRequest};
 use error::{Error, Result, SupError};
 use http_gateway;
